Moved the classes to module-aligned packages
