+++ caf-correlation/caf-correlation-id-constants/src/main/java/com/github/cafapi/correlation/constants/CorrelationIdConfigurationConstants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafapi.http.filters.correlationid;
+package com.github.cafapi.correlation.constants;
 
 public class CorrelationIdConfigurationConstants
 {
