+++ caf-correlation/caf-correlation-id-dropwizard/src/main/java/com/github/cafapi/correlation/dropwizard/CorrelationIdBundle.java
@@ -13,16 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.cafapi.http.filters.correlationid;
+package com.github.cafapi.correlation.dropwizard;
 
-import java.util.EnumSet;
-
-import javax.servlet.DispatcherType;
-
+import com.github.cafapi.correlation.jservlet.CorrelationIdFilter;
 import io.dropwizard.Configuration;
 import io.dropwizard.ConfiguredBundle;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
+import java.util.EnumSet;
+import javax.servlet.DispatcherType;
 
 public class CorrelationIdBundle<C extends Configuration> implements ConfiguredBundle<C> {
