+++ caf-correlation/caf-correlation-id-spring-tests/src/main/java/com/github/cafapi/correlation/spring/tests/GreetingController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.filter.http.spring;
+package com.github.cafapi.correlation.spring.tests;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
